Corrected an error in the definition of the Poisson Solver tolerance (parameter maxerror)
Gerris and Paris have now the same definition of the error. (Until now, there was a discrepancy
between Hypre and the standard solver in Paris.) i.e., it is the maximum magnitude of 
the relative change in volume of any cell in one timestep (i.e. the maximum error 
allowed in local volume/mass conservation).