Inflow boundary conditions
This patch allows to have inflow conditions on all faces, not only x- as previously. The uinject function modifies the inflow on the x- face. For this reason, a parameter 'Inject_type' is added in the parameter file. It is an integer. If it is set to 1 in the parameter file, it corresponds to a constant inflow velocity across the face. When set to 2, the old uinject function is retained. Note that the variable is initialized with value 2, meaning it will not affect cases where the uinject was used previously.