The pressure BC is improved, the old one is 5, and the new one is 6.
With Staggered grid, the pressure BC needs to be handled very carefully. 
It is challenging to statisfy both divergence free and the du/dn=0,
when the flow direction near the boundary is far from perpendicular. 
Therefore, it is difficult for a vortex to leave the boundary, 
extra attension is required when pressure BC is used.    

improve-pressure-BC.txt