add-HYPRE-GMRES-Solver.txt
The GMRES solver in the HYPRE library is also an option. The PFMG is used 
as a preconditioner. With number of preconditioning steps around 10, it is 
still much faster than the SMG solver, but nevertheless slower than the PFMG. 
Whether GMRES is more stable than PFMG is still not clear and some more 
tests are required.