Now the curvature is set to a parameter-defined value of UNCOMPUTED  when the computations fails or is impossible.
This minor patch is motivated as follows:  The curvature computation 
may fail for several reasons: not enough centroids, not enough mixed heights,
failure of the matrix inversion in the least square fit or the cell is far from the interface. 
When the computation failed, the curvature  used to be set to 0 in some of the cases and to 2d6
 in some other cases.This is somewhat inconsistent, so now the curvature is set to a parameter-defined value of 
"UNCOMPUTED" (in fact 2d6) when the computations fails or is impossible.