Detect singular matrices and in those cases declare parabolic fit failed.
This patch deals with the matrix inversion problem mentioned in the previous patch.  
If the final matrix has a diagonal element less than EPSC=1d-11 then it is considered
singular and the parabolic fit fails. This should ensure that when the parabolic fit does not
fail the curvature is determined with a good accuracy.

If the parabolic fit fails, the curvature is set to UNCOMPUTED and there are fractional
cells with no curvature, thus there are locations where there is no surface tension force, 
which is perhaps better than having a non-zero surface tension force with an O(1) error. 

Even without this patch, there were already many points with failed parabolic fits, 
but less of them, which resulted in no points with no surface tension force. 
Now there are statistically many more points with no surface tension.