Small VOF debris can now be removed after tagging and computing the total volume of the VOF strucure.
Debris here is by default defined as a VOF structure smaller than a cell, though a new input vof_debris
can be used a threshold. 

Note that You need to turn on the DoConvertVOF2LPP flag to enable this feature.

In the VOF-to-LPP conversion, now it is divided into two categories: 
1, small debris (e.g., vol<dx^3): simply removed, not stored or tracked, no undisturbed flow reconstruction
2, small droplets (including drop and drop_merge): converted to LPP, tracked in time, undisturbed flow field
is reconstructed during the conversion. 

The removal of small debris seems to be important in stablising simulations in particular 
when the momentum-conserving VOF is used.  


remove_debris_in_lpp-tagging.txt