Add debug mode functionality
+This patch affects all source files. Additional subroutine is added, deb_inf("name") which is a debugging-helper, by means
of automatic bracketting system. To enable it, put "debug_mode=T" into the input file. By default, the debug_mode is off.
Once enabled (by "debug_mode = T" in 'input' file) it turns Paris into a mode in subroutines print their names to the
screen when executed. While this is overkill (and doesn't make sense) in normal production use, it is a big help
once bugs appear as they become trivial to find quickly withot gdb and manual settings of gdb-breakpoints.
This mode is meant for preliminary bug-searching.
+additionally this patch does some super-small security fix in new_alpha3d