Correct problem causing a NaN result in a few of the vof_functions routines
The origin is that when the volume fraction is either 0 or 1 (which is
again set to zero by using the simmetry of the functions with respect
to cc = alpha = 1/2), and at least one of the normal coefficients is zero
we have results that are 0/0, even if the numerator goes to zero
quadratically and the denominator linearly.

To avoid this issue I have defined
ccr = DMAX1(cch,eps0)
where eps0 is a very small parameter, and in one routine this is done
for alh as well.

From Ruben Scardovelli.